Create Model Search Object Definitions

Functions for creating model type definitions for objects in the Feature Selection framework. Current available search types/methods are:

forward step-wise model

backward step-wise model

This information is called within the object creation to modify the its attributes (class) to reflect the appropriate search values.

Usage

search_type_backward_model(display_name = "Backward Stepwise Model Search")

search_type_forward_model(
  display_name = "Forward Stepwise Model Search",
  max_steps = 20L
)

Arguments

display_name

character(1). A title used by S3 plot methods.

max_steps

integer(1). Maximum number of features allowed in the model.

Value

A list containing:

display_name

The official "Title" to be used by downstream plot methods called on the feature_select object.

max_steps

Maximum model steps to search. Forward search only!

Functions

search_type_backward_model(): Backward model selection search.

search_type_forward_model(): Forward model selection search.

Author

Stu Field, Kirk DeLisle

Examples

search_type_forward_model()                    # the default = 20L
#> ── Forward Search ─────────────────────────────────────────────────────
#> • display_name  'Forward Stepwise Model Search'
#> • max_steps     20
#> ───────────────────────────────────────────────────────────────────────

search_type_forward_model(max_steps = 15L)     # set to 15
#> ── Forward Search ─────────────────────────────────────────────────────
#> • display_name  'Forward Stepwise Model Search'
#> • max_steps     15
#> ───────────────────────────────────────────────────────────────────────

search_type_forward_model("My Forward Search") # set a display title
#> ── Forward Search ─────────────────────────────────────────────────────
#> • display_name  'My Forward Search'
#> • max_steps     20
#> ───────────────────────────────────────────────────────────────────────