Set up Cross-Validation Folds

These S3 methods set up the internal cross-validation. Methods exist for both stratified and non-stratified scenarios. Since these are mostly internal methods, documentation is minimal.

The stratified version is a S3 method that sets up the internal cross-validation for stratified folds. Stratified implies that the structure (proportions) of the response variable is maintained (i.e. the proportion of disease samples in the fold is comparable to the disease proportion in the full training set).

The plotting routine allows one to visually check the cross-validation stratification for bias.

Usage

setup_cross(x)

setup_cross_strat(x)

plot_cross(x)

Arguments

x

A feature_select class object from a call call to feature_selection().

Author

Stu Field